requirements.txt
Babel==2.8.0
billiard==3.6.3.0
blinker==1.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
dnspython==2.0.0
email-validator==1.1.2
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-SQLAlchemy==2.4.4
python-decouple
gunicorn
idna==2.10
importlib-metadata==3.0.0
itsdangerous==1.1.0
MarkupSafe==1.1.1
passlib==1.7.4
permissive-dict==1.0.3
prompt-toolkit==3.0.8
PyJWT==1.7.1
psycopg2==2.8.6
psycopg2-binary==2.9.3
python-dotenv==0.15.0
python-decouple==3.3
pytz==2020.4
six==1.15.0
speaklater==1.3
SQLAlchemy==1.3.20
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.3
zipp==3.4.0
marshmallow-sqlalchemy==0.24.1
flask-marshmallow
flask-restplus
marshmallow~=3.15.0
config~=0.5.0.post0
Faker~=13.11.1
pytest~=7.1.2
decouple~=0.0.7